requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
tensorflow==2.19.0
numpy==1.26.4
pillow==10.2.0
python-multipart==0.0.9